Cache the output of the GeoJSON page; use Associationistrict broudary for map BB
Use Django's cache to store the content of the GeoJSON page
(which is slow to generate) "forever". Invalidaye the cache
on any save() or delete() of any of the database's model instances.

Also, use the extent to the Association boundary to set the BB of the map -
this is easily big enough, and loads quickly so at least a map appears even if the
tower markers are being slow.